requirements.txt
# Core Scientific Computing
numpy>=1.21.0
scipy>=1.8.0
pandas>=2.0.0

# Optimization & Convex
cvxpy>=1.3.0
mosek>=10.0.0
networkx>=3.0

# Visualization Tools
plotly>=5.17.0
matplotlib>=3.7.0
pyvista>=0.43.0
meshcat>=0.3.0
seaborn>=0.13.0

# Jupyter & Interactive
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
ipympl>=0.9.3
nbconvert>=7.0.0

# Training & Monitoring
wandb>=0.15.0
tensorboard>=2.14.0
tqdm>=4.66.0

# Image & Signal Processing
scikit-image>=0.21.0
pillow>=9.0.0

# Development Tools
black>=23.0.0
pylint>=2.17.0
flake8>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
